moving UI model ListItem to the model package
