+++ app/src/main/java/com/ivan/m/reddittimeline/model/ui/HomeUi.kt
@@ -1,7 +1,5 @@
 package com.ivan.m.reddittimeline.model.ui
 
-import com.ivan.m.reddittimeline.ui.home.ListItem
-
 enum class UiStatus {
     VALID, EMPTY, ERROR
 }
